Sky or Hell 2.0 - v0.48
Yo Gabriel hice todo xd
Vincule las notas MIDI del piano real con el piano virtual, funcionando tanto el teclado del pc como el teclado de piano real.

+++ Sky or Hell 2.0 - Apocalypse Piano/Assets/Scrips C#/Keyboard.cs
@@ -12,7 +12,7 @@ public class Keyboard : MonoBehaviour
     private Color myColor;*/
 
 
-    public int numberOfOctaves;
+    public float numberOfOctaves;
     // Start is called before the first frame update
     void Start()
     {
